+++ android/src/main/java/com/klaviyoreactnativesdk/KlaviyoReactNativeSdkPackage.kt
@@ -9,7 +9,6 @@ class KlaviyoReactNativeSdkPackage : ReactPackage {
   override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> =
     listOf(
       KlaviyoReactNativeSdkModule(reactContext),
-      KlaviyoFormsReactNativeSdkModule(reactContext),
     )
 
   override fun createViewManagers(reactContext: ReactApplicationContext): List<ViewManager<*, *>> = emptyList()
