--- ios/KlaviyoFormsReactNativeSdk-Bridging-Header.h
@@ -1,6 +1,0 @@
-// KlaviyoFormsReactNativeSdk-Bridging-Header.h
-#import <React/RCTBridgeModule.h>
-
-@interface KlaviyoFormsReactNativeSdk : NSObject <RCTBridgeModule>
-
-@end
