+++ src/Forms.ts
@@ -1,0 +1,10 @@
+/**
+ * Interface for the Klaviyo Forms API
+ */
+export interface KlaviyoFormsApi {
+  /**
+   * Load in-app forms data and display a form to the user if applicable based on the forms
+   * configured in your Klaviyo account. Note [Klaviyo.initialize] must be called first
+   */
+  registerForInAppForms(): void;
+}
