--- src/forms.tsx
@@ -1,11 +1,0 @@
-import { KlaviyoFormsReactNativeSdk } from './KlaviyoFormsReactNativeSdk';
-
-/**
- * Klaviyo Forms
- */
-
-export const KlaviyoForms = {
-  registerForInAppForms(): void {
-    KlaviyoFormsReactNativeSdk.registerForInAppForms();
-  },
-};
