+++ src/index.tsx
@@ -59,6 +59,9 @@ export const Klaviyo: KlaviyoInterface = {
   createEvent(event: Event): void {
     KlaviyoReactNativeSdk.createEvent(event);
   },
+  registerForInAppForms(): void {
+    KlaviyoReactNativeSdk.registerForInAppForms();
+  },
 };
 
 export { type Event, type EventProperties, EventName } from './Event';
@@ -70,4 +73,3 @@ export {
   ProfileProperty,
 } from './Profile';
 export type { KlaviyoInterface } from './Klaviyo';
-export { KlaviyoForms } from './forms';
